Initial Scraper Commit
Choose to use JavaScript because I thought it would be the best lang for the
job because of its relationship with the DOM. WebDriver abstracts everything
developer so JS isn't helping anything. But the bindings for JS are better than
the ones for go.

This html is so inefficient. There are random tags nested for no layout or
semantic reason.